Allow purchase price in posting
The code is pretty ugly, but at least it works. An idea for a future
improvement might be to tokenize within the string, with regular
expressions that take a given start of a string and the rest. For
example a regex like #"( @@ )(.*)" could be used to recognize that we're
at the start of a purchase price.
If all regexes have this form (ending with a group to catch 'the rest')
this could probably be used in a reduce like way to parse a string with
optional information.
Plus, it would allow for reuse of some regular expressions.