Rewrite parse-budget and parse-transaction
Use loop/recur. It already bothered me that I did explicit recursion
(although that might not eat up your stack as part of trampoline, but
not sure). This also improves the signature of parse-budget, where it's
weird that you have to supply the budget-accumulator at the first call.
Same for parse-transaction.