End of day: add parser based on instaparse
After some struggling, I finally arrived at a nice solution using
instaparse. Still some to do's (see comment), but it's the start of a
new parser.

The trouble with the finite state machine + regex approach is that you
can't do a look-ahead very well. The instaparse package handles all the
troubles of parsing the string into a tree based on the rules you
provide.

The read functions (I should check if the name is correct/common) can
then use destructuring based on the known structure of a node in the
resulting tree. It feels a bit as though I'm repeating the logic of the
structure of the file in both the parser definition and in the reader
functions. It might be inevitable, but I could also imagine that it's
possible to supply a function to the parser that it should apply to the
result. I guess I have some reading left to do :)